Move append_integer
